api/pkg/stdcopy: move to client/pkg/stdcopy
