+++ client/pkg/stdcopy/stdcopy.go
