+++ daemon/server/httputils/write_log_stream.go
@@ -8,7 +8,7 @@ import (
 	"net/url"
 	"sort"
 
-	"github.com/moby/moby/api/pkg/stdcopy"
+	"github.com/moby/moby/client/pkg/stdcopy"
 	"github.com/moby/moby/v2/daemon/internal/stdcopymux"
 	"github.com/moby/moby/v2/daemon/server/backend"
 	"github.com/moby/moby/v2/pkg/ioutils"
