+++ integration-cli/docker_api_attach_test.go
@@ -11,9 +11,9 @@ import (
 	"time"
 
 	"github.com/docker/go-connections/sockets"
-	"github.com/moby/moby/api/pkg/stdcopy"
 	"github.com/moby/moby/api/types"
 	"github.com/moby/moby/client"
+	"github.com/moby/moby/client/pkg/stdcopy"
 	"github.com/moby/moby/v2/integration-cli/cli"
 	"github.com/moby/moby/v2/internal/testutil"
 	"github.com/moby/moby/v2/internal/testutil/request"
